+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Make IntelliJ Platform dependencies available to `testImplementation` by default to avoid mutating tests classpath
 - For running testing, use test sandbox instead of compiled classes from `build/classes` directory 
 - Updated the `publishPlugin()` task to use IDE Services’ new `POST /api/plugins` API when configured to publish to the IDE Services plugin repository, replacing the deprecated `POST /api/ij-plugins/upload` endpoint
+- Rewrite bundled plugins/modules resolution by using IntelliJ Structure library.
 
 ### Fixed